requirements.txt
python-dotenv
streamlit
openai
fastapi
uvicorn
transformers
torch
groq
mistralai
soundfile
pillow
requests
pandas
python-multipart